+++ tests/test-cran/test-paths.R
@@ -7,7 +7,9 @@ assert('file_ext() and sans_ext() work', {
   (file_ext(c('foo.bar.gz', 'foo', 'file.nb.html')) %==% c('gz', '', 'nb.html'))
   # some special extensions
   p = c('abc.c++', 'def.c--', 'ghi.e##', 'jkl.FB2K-COMPONENT', 'mno.WITNESS_CAMPAIGN', 'pqr.H!')
-  (file_ext(p) %==% c('c++', 'c--', 'e##', 'FB2K-COMPONENT', 'WITNESS_CAMPAIGN', 'H!'))
+  (file_ext(p, '-+!_#') %==% c('c++', 'c--', 'e##', 'FB2K-COMPONENT', 'WITNESS_CAMPAIGN', 'H!'))
+  # by default, these extensions are not recognized
+  (file_ext(p) %==% character(length(p)))
 })
 
 assert('with_ext() works for corner cases', {